🎨 extract Http and GitHub in modules
